refactor: extract Inspector + object-window pane from WorkspaceLive BT-3291
Moves the docked Inspector's inspect/drill/crumb/freeze/poke navigation and
the BT-2493 floating-window overlay out of the 12k-line WorkspaceLive into a
new BtAttachWeb.Live.Inspector module, so those branches (field-flash
coalescing, freeze/unfreeze, drill/crumb navigation, the owner-only poke RBAC
gate, window position reset) are directly unit-testable instead of reachable
only through the :workspace-tagged full-stack WorkspaceLiveTest.

WorkspaceLive.handle_event/3 and handle_info/2 now delegate to Inspector by
event/message name; the floating-window function components move with their
logic. Two small pieces of duplicated logic (ctx/1's RBAC context builder and
the facade-error renderer) were extracted into shared leaf modules
(RequestContext, FacadeError) so WorkspaceLive and Inspector share one copy
rather than each keeping its own. The `<.notice>` dismissable-banner
component moved from a WorkspaceLive-private component into CoreComponents
(already auto-imported everywhere) for the same reason.

workspace_live.ex drops from 12318 to 10686 lines and its handle_event
clause count from 171 to 144. Added 28 direct unit tests for Inspector
against a hand-built socket + the stubbed workspace client, raising its
coverage from ~10% to ~49% (project total 59.85% -> 64.23%, above the 55%
threshold).
